Move resourcekey.go to a separate package
